+++ crates/iroha_core/src/smartcontracts/isi/asset.rs
@@ -10,23 +10,6 @@ use iroha_telemetry::metrics;
 
 use super::prelude::*;
 
-impl Registrable for NewAssetDefinition {
-    type Target = AssetDefinition;
-
-    #[inline]
-    fn build(self, authority: &AccountId) -> Self::Target {
-        Self::Target {
-            id: self.id,
-            spec: self.spec,
-            mintable: self.mintable,
-            logo: self.logo,
-            metadata: self.metadata,
-            owned_by: authority.clone(),
-            total_quantity: Numeric::ZERO,
-        }
-    }
-}
-
 /// ISI module contains all instructions related to assets:
 /// - minting/burning assets
 /// - update metadata
